refactored operation logs. added unique index to dataset versions
