+++ core/src/models/operation_logs.rs
@@ -11,29 +11,10 @@ use serde::{Deserialize, Serialize};
 use strum::Display;
 use uuid::Uuid;
 
+use super::logs::{Action, LogOperation, LogOperationDataset};
 use super::{schema, Dataset, DatasetVersion, PublicationType, TaxonomicRank, TaxonomicStatus};
 use crate::crdt::DataFrameOperation;
 use crate::models::NomenclaturalActType;
-
-#[derive(Clone, Debug, Display, Serialize, Deserialize, diesel_derive_enum::DbEnum)]
-#[ExistingTypePath = "schema::sql_types::OperationAction"]
-pub enum Action {
-    Create,
-    Update,
-}
-
-pub trait LogOperation<T> {
-    /// The hash of the entity id
-    fn id(&self) -> &BigDecimal;
-    fn entity_id(&self) -> &String;
-    fn action(&self) -> &Action;
-    fn atom(&self) -> &T;
-}
-
-pub trait LogOperationDataset {
-    fn dataset_version(&self) -> &DatasetVersion;
-    fn dataset(&self) -> &Dataset;
-}
 
 
 #[derive(Atom, Debug, Default, Clone, Serialize, Deserialize, AsExpression, FromSqlRow, PartialEq, Display)]
@@ -240,167 +221,6 @@ pub struct NomenclaturalActOperation {
 }
 
 
-#[derive(Atom, Debug, Default, Clone, Serialize, Deserialize, AsExpression, FromSqlRow, PartialEq, Display)]
-#[diesel(sql_type = diesel::sql_types::Jsonb)]
-pub enum SpecimenAtom {
-    #[default]
-    Empty,
-
-    /// The globally unique specimen id.
-    SpecimenId(String),
-
-    /// Used to link the collection to a name.
-    ScientificName(String),
-
-    /// When the specimen registration happened. Strictly YYYY-MM-DD
-    EventDate(String),
-    /// What time the specimen registration happened. Strictly HH:MM:SS
-    EventTime(String),
-
-    /// The name of the institution that owns the specimen.
-    InstitutionName(String),
-    /// The short code of the institution.
-    InstitutionCode(String),
-    /// The ID for the specific specimen. Typically the voucher registration number.
-    CollectionRepositoryId(String),
-    /// The code for the specific collection repository in the institution.
-    CollectionRepositoryCode(String),
-
-    TypeStatus(String),
-
-    Preparation(String),
-
-    OtherCatalogNumbers(String),
-
-    Disposition(String),
-}
-
-#[derive(OperationLog, Queryable, Selectable, Insertable, Associations, Debug, Serialize, Deserialize, Clone)]
-#[diesel(belongs_to(DatasetVersion))]
-#[diesel(table_name = schema::specimen_logs)]
-#[diesel(check_for_backend(diesel::pg::Pg))]
-pub struct SpecimenOperation {
-    pub operation_id: BigDecimal,
-    pub parent_id: BigDecimal,
-    pub entity_id: String,
-    pub dataset_version_id: Uuid,
-    pub action: Action,
-    pub atom: SpecimenAtom,
-}
-
-
-#[derive(Atom, Debug, Clone, Default, Serialize, Deserialize, AsExpression, FromSqlRow, PartialEq, Display)]
-#[diesel(sql_type = diesel::sql_types::Jsonb)]
-pub enum CollectionEventAtom {
-    #[default]
-    Empty,
-    /// The globally unique collection id.
-    FieldCollectingId(String),
-    /// The global specimen id of the collected specimen if any.
-    SpecimenId(String),
-    /// Used to link the collection to a name
-    ScientificName(String),
-
-    /// When the collection happened. Strictly YYYY-MM-DD
-    EventDate(chrono::NaiveDate),
-    /// What time the collection happened. Strictly HH:MM:SS
-    EventTime(chrono::NaiveTime),
-
-    /// The name of the person who did the collection.
-    CollectedBy(String),
-    /// Free-text notes about the collection event.
-    CollectionRemarks(String),
-
-    /// The name of the person who identified the organism at collection.
-    IdentifiedBy(String),
-    /// The date the organism collection was identified. Strictly YYYY-MM-DD.
-    IdentifiedDate(chrono::NaiveDate),
-    /// Free-text notes about the identification of the collection event.
-    IdentificationRemarks(String),
-
-    /// The global identifier for the organism that was collected
-    OrganismId(String),
-
-    Locality(String),
-    Country(String),
-    CountryCode(String),
-    StateProvince(String),
-    County(String),
-    Municipality(String),
-    Latitude(f64),
-    Longitude(f64),
-    Elevation(f64),
-    Depth(f64),
-    ElevationAccuracy(f64),
-    DepthAccuracy(f64),
-    LocationSource(String),
-
-    Preparation(String),
-
-    EnvironmentBroadScale(String),
-    EnvironmentLocalScale(String),
-    EnvironmentMedium(String),
-
-    /// The habitat this collection was made in.
-    Habitat(String),
-
-    /// Scientific name of the host species.
-    SpecificHost(String),
-
-    IndividualCount(String),
-    OrganismQuantity(String),
-    OrganismQuantityType(String),
-
-    Strain(String),
-    Isolate(String),
-    FieldNotes(String),
-}
-
-
-#[derive(OperationLog, Queryable, Selectable, Insertable, Associations, Debug, Serialize, Deserialize, Clone)]
-#[diesel(belongs_to(DatasetVersion))]
-#[diesel(table_name = schema::collection_event_logs)]
-#[diesel(check_for_backend(diesel::pg::Pg))]
-pub struct CollectionEventOperation {
-    pub operation_id: BigDecimal,
-    pub parent_id: BigDecimal,
-    pub entity_id: String,
-    pub dataset_version_id: Uuid,
-    pub action: Action,
-    pub atom: CollectionEventAtom,
-}
-
-
-#[derive(Atom, Debug, Clone, Default, Serialize, Deserialize, AsExpression, FromSqlRow, PartialEq, Display)]
-#[diesel(sql_type = diesel::sql_types::Jsonb)]
-pub enum OrganismAtom {
-    #[default]
-    Empty,
-
-    /// Used to link the organism to a name
-    ScientificName(String),
-
-    Sex(String),
-    GenotypicSex(String),
-    PhenotypicSex(String),
-    LifeStage(String),
-    ReproductiveCondition(String),
-    Behavior(String),
-}
-
-#[derive(OperationLog, Queryable, Selectable, Insertable, Associations, Debug, Serialize, Deserialize, Clone)]
-#[diesel(belongs_to(DatasetVersion))]
-#[diesel(table_name = schema::organism_logs)]
-#[diesel(check_for_backend(diesel::pg::Pg))]
-pub struct OrganismOperation {
-    pub operation_id: BigDecimal,
-    pub parent_id: BigDecimal,
-    pub entity_id: String,
-    pub dataset_version_id: Uuid,
-    pub action: Action,
-    pub atom: CollectionEventAtom,
-}
-
 #[derive(Atom, Debug, Clone, Default, Serialize, Deserialize, AsExpression, FromSqlRow, PartialEq, Display)]
 #[diesel(sql_type = diesel::sql_types::Jsonb)]
 pub enum TaxonDistributionAtom {
@@ -483,4 +303,53 @@ pub struct PublicationOperation {
     pub dataset_version_id: Uuid,
     pub action: Action,
     pub atom: PublicationAtom,
+}
+
+
+#[derive(Atom, Debug, Default, Clone, Serialize, Deserialize, AsExpression, FromSqlRow, PartialEq, Display)]
+#[diesel(sql_type = diesel::sql_types::Jsonb)]
+pub enum SpecimenAtom {
+    #[default]
+    Empty,
+
+    /// The globally unique specimen id.
+    SpecimenId(String),
+
+    /// Used to link the collection to a name.
+    ScientificName(String),
+
+    /// When the specimen registration happened. Strictly YYYY-MM-DD
+    EventDate(String),
+    /// What time the specimen registration happened. Strictly HH:MM:SS
+    EventTime(String),
+
+    /// The name of the institution that owns the specimen.
+    InstitutionName(String),
+    /// The short code of the institution.
+    InstitutionCode(String),
+    /// The ID for the specific specimen. Typically the voucher registration number.
+    CollectionRepositoryId(String),
+    /// The code for the specific collection repository in the institution.
+    CollectionRepositoryCode(String),
+
+    TypeStatus(String),
+
+    Preparation(String),
+
+    OtherCatalogNumbers(String),
+
+    Disposition(String),
+}
+
+#[derive(OperationLog, Queryable, Selectable, Insertable, Associations, Debug, Serialize, Deserialize, Clone)]
+#[diesel(belongs_to(DatasetVersion))]
+#[diesel(table_name = schema::specimen_logs)]
+#[diesel(check_for_backend(diesel::pg::Pg))]
+pub struct SpecimenOperation {
+    pub operation_id: BigDecimal,
+    pub parent_id: BigDecimal,
+    pub entity_id: String,
+    pub dataset_version_id: Uuid,
+    pub action: Action,
+    pub atom: SpecimenAtom,
 }
